base requirements met for upload

+++ flaskr/backend.py
@@ -1,63 +1,77 @@
 # TODO(Project 1): Implement Backend according to the requirements.
-# Imports the Google Cloud client library
 from google.cloud import storage
 
 
 class Backend:
-
     def __init__(self, opener = open):
-        self.writeBucket =opener
-        self.readBucket =opener
         self.storage_client = storage.Client()
+
         
         self.bucketName_content = "group_wiki_content"
         self.bucketName_users = "users_and_passwords"
 
-        #Creating the buckets
-        #The bucket already exists, it just needs a name
+
         self.bucket_content = self.storage_client.bucket(self.bucketName_content)
-        self.bucket_users = self.storage_client.bucket(self.bucketName_users)
+        self.bucket_users = self.storage_client.bucket(self.bucketName_users)      
         
     def get_wiki_page(self, name):
-        blob = self.bucket_content.blob(self.bucketName_content+"/"+name)
+        blob = self.bucket_content.blob(name)
         with blob.open("r") as f:
-            print(f.read())
+            return (f.read())
+        
 
     def get_all_page_names(self):
-        blobs = self.bucket_content.list_blobs(self.bucketName_content)        
+        pages = []
+        blobs = self.storage_client.list_blobs(self.bucketName_content) 
         for blob in blobs:
-            print(blob.name)
+            pages.append(blob.name)
+        return pages
 
-    def upload(self, content, name):  
+    def upload(self, content, name):
         """Uploads content into the group_wiki_content bucket
 
         It should be able to store text and pictures
         Attributes
-            content: 
-        """  
-        #The blob will be the name of the object the bucket will use to identify it
-        blob = self.bucket_content.blob(self.bucketName_content+"/"+name)
-
-        print(f"Bucket {self.bucketName_content} created.")
+            content: the wikipage
+            name:  the id of the wikipage 
+        """          
+        blob = self.bucket_content.blob(name)
         with blob.open("w") as f:
-            print(f"Bucket {self.bucketName_content} created.")
             f.write(content)        
-        print(f"Bucket {self.bucketName_content} created.")
 
-    def sign_up(self, username, password):
-        blob = self.bucket_users.blob(self.bucketName_users+"/"+username)
+
+
+    def sign_up(self, username, password):        
+        if self.user_exists(username):
+            return False
+
+        blob = self.bucket_users.blob(username)
         with blob.open("w") as f:
             f.write(password)
+        return True
+
+    def user_exists(self, username):
+        blobs = self.storage_client.list_blobs(self.bucketName_users) 
+        for blob in blobs:
+            if blob.name == username:
+                return True
+        return False
+
 
     def sign_in(self, username, password):
-        blob = self.bucket_content.blob(self.bucketName_content+"/"+username)
-        matched = False
-        with blob.open("r") as f:
-            retrievedPassword = f.read()
-            print(retrievedPassword)
-            if retrievedPassword is password:
-                matched = True
-        return matched
+            
+        if not self.user_exists(username):
+            return False
+
+        blobContent = self.bucket_users.blob(username) 
+
+        #compare passwords
+        with blobContent.open("r") as f:
+            blobPassword = f.read()
+            for line in blobPassword:
+                if blobPassword == password:
+                    return True
+        return False
 
     def get_image(self, imageName):
         blob = self.bucket_content.blob(self.bucketName_content+"/"+imageName)
@@ -65,4 +79,6 @@ class Backend:
         with blob.open("r") as f:
             image = f.read()
         return image
+
+
